is it really possible to be this much smaller?
