+++ package.json
@@ -9,9 +9,10 @@
     "url": "git://github.com/nlp-compromise/efrt.git"
   },
   "scripts": {
-    "test": "node ./scripts/test.js",
-    "testb": "TESTENV=prod node ./scripts/test.js",
-    "build": "node ./scripts/build/index.js",
+    "build": "rollup -c && node ./scripts/filesize.js",
+    "filesize": "node ./scripts/filesize.js",
+    "test": "tape \"./tests/*.test.js\" | tap-dancer",
+    "testb": "TESTENV=prod tape \"./tests/*.test.js\" | tap-dancer",
     "watch": "amble ./scratch"
   },
   "files": [
@@ -29,17 +30,13 @@
     "@babel/core": "7.5.5",
     "@babel/preset-env": "7.5.5",
     "amble": "0.0.7",
-    "babelify": "10.0.0",
-    "browserify": "16.3.0",
-    "codacy-coverage": "3.4.0",
-    "derequire": "^2.0.3",
-    "eslint": "6.0.1",
-    "gaze": "1.1.3",
-    "nyc": "14.1.1",
-    "shelljs": "0.8.3",
     "tap-spec": "5.0.0",
     "tape": "4.11.0",
-    "terser": "^4.0.0"
+    "rollup-plugin-babel": "^4.3.3",
+    "rollup-plugin-commonjs": "10.0.1",
+    "rollup-plugin-json": "^4.0.0",
+    "rollup-plugin-node-resolve": "5.2.0",
+    "rollup-plugin-terser": "5.1.1"
   },
   "license": "MIT"
 }
