+++ scripts/filesize.js
@@ -1,34 +1,16 @@
-//test the compression amount
-require('shelljs/global')
 const fs = require('fs')
-const efrt = require('../src')
+const path = require('path')
 
-const pos = 'nouns'
-console.log('===' + pos + '===')
-const file = '/home/spencer/nlp/wordnet.js/' + pos + '.json'
-const out = './tmp.txt'
+console.log('\n\n----\n')
+const buildsDir = path.join(__dirname, '../builds')
 
-const fileSize = function(src) {
-  const stats = fs.statSync(src)
-  const num = (stats['size'] / 1000.0).toFixed(2)
-  console.log('  ' + num + 'kb')
-  return num
+const doFile = function(file) {
+  const abs = path.join(buildsDir, file)
+  const stats = fs.statSync(abs)
+  const size = (stats['size'] / 1000.0).toFixed(1)
+  console.log('    ' + file + '  : ' + size + ' kb')
+  console.log('\n')
 }
 
-const arr = require(file)
-console.log('before:')
-console.log('  ' + arr.length + ' words')
-const before = fileSize(file)
-console.log('  ...\n\n')
-const str = efrt.pack(arr)
-
-console.log('after:')
-fs.writeFileSync(out, str)
-const after = fileSize(out)
-
-console.log('\n\n')
-
-const percent = (before - after) / before
-console.log(parseInt(percent * 100) + '% compressed')
-
-exec('rm ' + out)
+doFile('./efrt.min.js')
+doFile('./efrt-unpack.min.js')
