+++ tests/data/adjectives.json
