+++ tests/lib/efrt.js
